+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/ParallelStreamProcessor.java
@@ -4,6 +4,7 @@ package io.confluent.parallelconsumer;
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
+import io.confluent.parallelconsumer.internal.DrainingCloseable;
 import lombok.Data;
 import lombok.SneakyThrows;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
