+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/RateLimiter.java
@@ -1,4 +1,4 @@
-package io.confluent.parallelconsumer;
+package io.confluent.parallelconsumer.internal;
 
 /*-
  * Copyright (C) 2020-2021 Confluent, Inc.
