+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/UserFunctions.java
@@ -1,9 +1,11 @@
-package io.confluent.parallelconsumer;
+package io.confluent.parallelconsumer.internal;
 
 /*-
  * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
+
+import io.confluent.parallelconsumer.ErrorInUserFunctionException;
 import lombok.experimental.UtilityClass;
 
 import java.util.function.BiFunction;
