+++ parallel-consumer-core/src/test/java/io/confluent/csid/utils/KafkaTestUtils.java
@@ -1,12 +1,12 @@
 package io.confluent.csid.utils;
 
 /*-
- * Copyright (C) 2020 Confluent, Inc.
+ * Copyright (C) 2020-2021 Confluent, Inc.
  */
 
 import io.confluent.parallelconsumer.OffsetMapCodecManager;
-import io.confluent.parallelconsumer.WorkContainer;
-import io.confluent.parallelconsumer.WorkManager;
+import io.confluent.parallelconsumer.state.WorkContainer;
+import io.confluent.parallelconsumer.state.WorkManager;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
