+++ parallel-consumer-core/src/test/resources/logback-test.xml
@@ -30,7 +30,7 @@
     <!--    <logger name="io.confluent.parallelconsumer" level="debug"/>-->
     <!--        <logger name="io.confluent.parallelconsumer" level="trace"/>-->
     <!--            <logger name="io.confluent.parallelconsumer.ParallelEoSStreamProcessor" level="trace"/>-->
-    <!--                <logger name="io.confluent.parallelconsumer.BrokerPollSystem" level="trace"/>-->
+    <!--                <logger name="io.confluent.parallelconsumer.internal.BrokerPollSystem" level="trace"/>-->
     <!--        <logger name="io.confluent.parallelconsumer.integrationTests.BrokerIntegrationTest" level="info"/> &lt;!&ndash; docker logs &ndash;&gt;-->
     <!--    <logger name="io.confluent.csid" level="info"/>-->
     <!--    <logger name="io.confluent.csid.utils" level="debug"/>-->
@@ -43,7 +43,7 @@
     <!--    <logger name="io.confluent.parallelconsumer.integrationTests" level="debug"/>-->
     <!--    <logger name="io.confluent.parallelconsumer.integrationTests.LargeVolumeInMemoryTests" level="debug"/>-->
     <!--    <logger name="io.confluent.parallelconsumer.ParallelConsumerTestBase" level="trace"/>-->
-    <!--        <logger name="io.confluent.parallelconsumer.WorkManager" level="trace" />-->
+    <!--        <logger name="io.confluent.parallelconsumer.state.WorkManager" level="trace" />-->
     <!--    <logger name="io.confluent.parallelconsumer.ParallelConsumer" level="info" />-->
     <!--    <logger name="io.confluent.parallelconsumer.OffsetMapCodecManager" level="debug"/>-->
     <!--    <logger name="io.confluent.parallelconsumer.WorkManagerOffsetMapCodecManagerTest" level="debug"/>-->
